removed default focus request from edit text
fixed the clear button to not just remove the holidays in shared prefs
but to clear the local datastructure and to call set date on the
calendar view to force a redraw. NEED TO LOOK INTO WHETHER THERE IS A
BETTER WAY TO DO THIS.

fixed the holiday decorator to standardize the coloring of dates between
the two calendar activities while letting the study coord select dates
outside of the valid date range to lable with holidays.