Modified export to export from a selected date to the most current.
edited the layout to have a dropdown of all dates on which data was
entered. Hopefully it is ordered correctly. neet to test. and a button
to actually do the export.

updated the export class to use the new layout. populate the dropdown
with the available dates.

updated the dbhelper getAllEntries method to take an optional Time
parameter to define the earliest date for which data should be returned.

updated the summarazeData class because the getAllEntries method
signature has changed.